+++ test/include/mock/mock_lsg_server.hpp
@@ -10,6 +10,7 @@ using ssize_t = SSIZE_T;
 #include <sys/socket.h>  // send
 #endif
 
+#include <chrono>
 #include <condition_variable>
 #include <functional>  // function
 #include <mutex>
@@ -28,16 +29,20 @@ class MockLsgServer {
  public:
   MockLsgServer(std::string dataset, bool ts_out,
                 std::function<void(MockLsgServer&)> serve_fn);
+  MockLsgServer(std::string dataset, bool ts_out,
+                std::chrono::seconds heartbeat_interval,
+                std::function<void(MockLsgServer&)> serve_fn);
 
   std::uint16_t Port() const { return port_; }
 
   void Accept();
   void Authenticate();
   void Subscribe(const std::vector<std::string>& symbols, Schema schema,
-                 SType stype, bool use_snapshot = false);
+                 SType stype);
   void Subscribe(const std::vector<std::string>& symbols, Schema schema,
-                 SType stype, const std::string& start,
-                 bool use_snapshot = false);
+                 SType stype, const std::string& start);
+  void SubscribeWithSnapshot(const std::vector<std::string>& symbols,
+                             Schema schema, SType stype);
   void Start();
   std::size_t Send(const std::string& msg);
   ::ssize_t UncheckedSend(const std::string& msg);
@@ -82,6 +87,7 @@ class MockLsgServer {
 
   std::string dataset_;
   bool ts_out_;
+  std::chrono::seconds heartbeat_interval_;
   std::uint16_t port_{};
   detail::ScopedFd socket_{};
   detail::ScopedFd conn_fd_{};