+++ test/src/mock_lsg_server.cpp
@@ -20,8 +20,14 @@ using databento::test::mock::MockLsgServer;
 
 MockLsgServer::MockLsgServer(std::string dataset, bool ts_out,
                              std::function<void(MockLsgServer&)> serve_fn)
+    : MockLsgServer{std::move(dataset), ts_out, {}, std::move(serve_fn)} {}
+
+MockLsgServer::MockLsgServer(std::string dataset, bool ts_out,
+                             std::chrono::seconds heartbeat_interval,
+                             std::function<void(MockLsgServer&)> serve_fn)
     : dataset_{std::move(dataset)},
       ts_out_{ts_out},
+      heartbeat_interval_{heartbeat_interval},
       socket_{InitSocketAndSetPort()},
       thread_{std::move(serve_fn), std::ref(*this)} {}
 
@@ -87,17 +93,39 @@ void MockLsgServer::Authenticate() {
   EXPECT_NE(received.find("encoding=dbn"), std::string::npos);
   EXPECT_NE(received.find("ts_out=" + std::to_string(ts_out_)),
             std::string::npos);
+  if (heartbeat_interval_.count() > 0) {
+    EXPECT_NE(received.find("heartbeat_interval_s=" +
+                            std::to_string(heartbeat_interval_.count())),
+              std::string::npos);
+  } else {
+    EXPECT_EQ(received.find("heartbeat_interval_s="), std::string::npos);
+  }
   Send("success=1|session_id=5|\n");
 }
 
 void MockLsgServer::Subscribe(const std::vector<std::string>& symbols,
-                              Schema schema, SType stype, bool use_snapshot) {
-  Subscribe(symbols, schema, stype, {}, use_snapshot);
+                              Schema schema, SType stype) {
+  Subscribe(symbols, schema, stype, "");
+}
+
+void MockLsgServer::SubscribeWithSnapshot(
+    const std::vector<std::string>& symbols, Schema schema, SType stype) {
+  const auto received = Receive();
+  EXPECT_NE(
+      received.find("symbols=" +
+                    JoinSymbolStrings("MockLsgServer::Subscribe", symbols)),
+      std::string::npos);
+  EXPECT_NE(received.find(std::string{"schema="} + ToString(schema)),
+            std::string::npos);
+  EXPECT_NE(received.find(std::string{"stype_in="} + ToString(stype)),
+            std::string::npos);
+  EXPECT_EQ(received.find("start="), std::string::npos);
+  EXPECT_NE(received.find("snapshot=1"), std::string::npos);
 }
 
 void MockLsgServer::Subscribe(const std::vector<std::string>& symbols,
                               Schema schema, SType stype,
-                              const std::string& start, bool use_snapshot) {
+                              const std::string& start) {
   const auto received = Receive();
   EXPECT_NE(
       received.find("symbols=" +
@@ -108,14 +136,11 @@ void MockLsgServer::Subscribe(const std::vector<std::string>& symbols,
   EXPECT_NE(received.find(std::string{"stype_in="} + ToString(stype)),
             std::string::npos);
   if (start.empty()) {
-    EXPECT_EQ(received.find(std::string{"start="}), std::string::npos);
+    EXPECT_EQ(received.find("start="), std::string::npos);
   } else {
     EXPECT_NE(received.find(std::string{"start="} + start), std::string::npos);
   }
-
-  EXPECT_NE(
-      received.find(std::string{"snapshot="} + std::to_string(use_snapshot)),
-      std::string::npos);
+  EXPECT_NE(received.find("snapshot=0"), std::string::npos);
 }
 
 void MockLsgServer::Start() {